Version 0.2
I am trying to add music but my data folder is not working. The
collisions now work and when it hits it s get transferred to another
screen. The image that I tried to use does not work as the data folder
is not working. My basic project is done but need to edit and fix a few
things.